chore: move all stats folders into a single `stats` folder
