+++ src/services/stats/socialLinkStats/socialLinkStats.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vite-plus/test';
 
 import { Arcanas } from '@constants/arcanas';
-import { SocialLinkLevel } from '@services/socialLinkLevel/socialLinkLevel';
+import { SocialLinkLevel } from '@services/stats/socialLinkLevel/socialLinkLevel';
 
 import { SocialLinkStats } from './socialLinkStats';
